"use_entity" and "render_to" support
This change allows any target entity to use another entity as a base,
picking up all keys which are not set in the target. (This enables
us to do subpages, a set of HTML output linked by a common subject.
Additionally, we add support for "render_to", enabling us to override
the default output for a given entity.